+++ src/util/type.ts
